Add support for temporarily removing roles from a user

+++ cogs/temporary_role_handler.py
@@ -21,13 +21,14 @@ class TemporaryRoleHandler(utils.Cog):
 
         # Get data
         async with self.bot.database() as db:
-            temporary_roles_rows = await db("SELECT * FROM temporary_roles")
-        if not temporary_roles_rows:
+            temporary_added_rows = await db("SELECT * FROM temporary_roles")  # For roles that may need removing
+            temporary_removed_rows = await db("SELECT * FROM temporary_removed_roles")  # For roles that may need readding
+        if not temporary_added_rows:
             return
 
-        # Remove roles
+        # Remove roles that have expired
         removed_roles = []
-        for row in temporary_roles_rows:
+        for row in temporary_added_rows:
             if row['remove_timestamp'] > dt.utcnow():
                 continue
 
@@ -59,13 +60,45 @@ class TemporaryRoleHandler(utils.Cog):
                 except (discord.Forbidden, discord.NotFound):
                     self.logger.info(f"Couldn't send DM to user about expired role (G{guild.id}/R{role.id}/U{member.id})")
 
+        # Readd roles that may have been removed
+        readded_roles = []
+        for row in temporary_removed_rows:
+            if row['readd_timestamp'] > dt.utcnow():
+                continue
+
+            # Get the relevant information
+            readded_roles.append((row['guild_id'], row['role_id'], row['user_id']))
+            guild = self.bot.get_guild(row['guild_id'])
+            member = guild.get_member(row['user_id'])
+            role = guild.get_role(row['role_id'])
+
+            # Remove the role
+            if role is not None and member is not None:
+                try:
+                    await member.remove_roles(role, reason="Role duration expired")
+                    self.logger.info(f"Removed role from user - duration expired (G{guild.id}/R{role.id}/U{member.id})")
+                except (discord.Forbidden, discord.NotFound) as e:
+                    self.logger.info(f"Couldn't remove duration expired role form user (G{guild.id}/R{role.id}/U{member.id}) - {e}")
+
+            # DM the user
+            if role is not None and member is not None and row['dm_user']:
+                try:
+                    await member.send(f"Added the `{role.name}` role from you in the server **{guild.name}** - removal duration expired.")
+                    self.logger.info(f"Sent DM to user about expired removed role (G{guild.id}/R{role.id}/U{member.id})")
+                except (discord.Forbidden, discord.NotFound):
+                    self.logger.info(f"Couldn't send DM to user about expired removed role (G{guild.id}/R{role.id}/U{member.id})")
+
         # Remove from db
-        async with self.bot.database() as db:
-            for guild_id, role_id, user_id in removed_roles:
-                await db("DELETE FROM temporary_roles WHERE guild_id=$1 AND role_id=$2 AND user_id=$3", guild_id, role_id, user_id)
+        if removed_roles or readded_roles:
+            async with self.bot.database() as db:
+                for guild_id, role_id, user_id in removed_roles:
+                    await db("DELETE FROM temporary_roles WHERE guild_id=$1 AND role_id=$2 AND user_id=$3", guild_id, role_id, user_id)
+                for guild_id, role_id, user_id in readded_roles:
+                    await db("DELETE FROM temporary_removed_roles WHERE guild_id=$1 AND role_id=$2 AND user_id=$3", guild_id, role_id, user_id)
 
         # Disconnect from db
         self.logger.info(f"Removed/deleted {len(removed_roles)} expired temporary roles")
+        self.logger.info(f"Added back {len(readded_roles)} expired temporary roles")
 
     @role_handler.before_loop
     async def before_role_handler(self):
